+++ src/device/tsp/Tsp/Script/IScriptResourceManager.cs
@@ -1,0 +1,122 @@
+using System.Reflection;
+
+namespace cc.isr.VI.Tsp.K2600.Ttmware;
+
+/// <summary>   Manager for resources. </summary>
+/// <remarks>   2025-04-05. </remarks>
+public interface IScriptResourceManager
+{
+    #region " embedded resource assembly info "
+
+    /// <summary>   Gets or sets the pathname of the embedded resource folder. </summary>
+    /// <value> The pathname of the embedded resource folder. </value>
+    public string EmbeddedResourceFolderName { get; set; }
+
+    /// <summary>   Gets or sets the identifier of the script framework. </summary>
+    /// <value> The identifier of the script framework. </value>
+    public string ScriptFrameworkId { get; set; }
+
+
+    /// <summary>   Gets or sets the assembly namespace. </summary>
+    /// <value> The assembly namespace. </value>
+    public string AssemblyNamespace { get; set; }
+
+    #endregion
+
+    #region " read embedded resource "
+
+    /// <summary>   Builds full resource name. </summary>
+    /// <remarks>   2025-04-08. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   A string. </returns>
+    public string BuildFullResourceName( string resourceFileName );
+
+    /// <summary>   Gets resource stream. </summary>
+    /// <remarks>   2025-04-07. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   The resource stream. </returns>
+    public Stream? GetResourceStream( string resourceFileName );
+
+    /// <summary>   Gets resource information. </summary>
+    /// <remarks>   2025-04-17. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   The resource information. </returns>
+    public ManifestResourceInfo? GetResourceInfo( string resourceFileName );
+
+    /// <summary>   Query if embedded resource <paramref name="resourceFileName"/> exists. </summary>
+    /// <remarks>   2025-04-17. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>
+    /// A tuple; True if resource exists, false if not, the resource full name, the assembly location.
+    /// </returns>
+    public (bool exists, string resourceFullName, string assemlyLocation) ResourceExists( string resourceFileName );
+
+    /// <summary>   Gets resource reader. </summary>
+    /// <remarks>   2025-04-15. <para>
+    /// Returning the stream reader allows rewinding. </para> </remarks>
+    /// <exception cref="FileNotFoundException">    Thrown when the requested file is not present. </exception>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   The resource <see cref="StreamReader"/> reader. </returns>
+    public StreamReader? GetResourceReader( string resourceFileName );
+
+    /// <summary>   Rewind stream reader. </summary>
+    /// <remarks>   2025-04-15. </remarks>
+    /// <exception cref="InvalidOperationException">    Thrown when the requested operation is
+    ///                                                 invalid. </exception>
+    /// <param name="reader">   The reader. </param>
+    /// <returns>   A StreamReader. </returns>
+    public StreamReader RewindStreamReader( StreamReader reader );
+
+    /// <summary>   Reads an embedded resource given the full resource path. </summary>
+    /// <remarks>   2025-04-01. </remarks>
+    /// <exception cref="FileNotFoundException">    Thrown when the requested file is not present. </exception>
+    /// <param name="resourceName"> Name of the resource. </param>
+    /// <returns>   The embedded resource content. </returns>
+    public string ReadEmbeddedResourceFromFullName( string resourceName );
+
+    /// <summary>   Reads embedded resource. </summary>
+    /// <remarks>   2025-04-03. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   The embedded resource. </returns>
+    public string ReadEmbeddedResource( string resourceFileName );
+
+    #endregion
+
+    #region " read resource file "
+
+    /// <summary>   Builds the resource file folder path. </summary>
+    /// <remarks>   2025-04-22. </remarks>
+    /// <param name="resourceFolderName">   (Optional) ['resources'] name of the folder holding the
+    ///                                     resource. </param>
+    /// <returns>   A string. </returns>
+    public string BuildResourceFileFolderPath( string resourceFolderName = "resources" );
+
+    /// <summary>   Builds full resource file path. </summary>
+    /// <remarks>   2025-04-08. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   A string. </returns>
+    public string BuildResourceFilePath( string resourceFileName );
+
+    /// <summary>   Gets resource file info. </summary>
+    /// <remarks>   2025-04-07. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   The resource file info. </returns>
+    public FileInfo? GetResourceFileInfo( string resourceFileName );
+
+    /// <summary>   Reads resource file. </summary>
+    /// <remarks>   2025-04-07. </remarks>
+    /// <exception cref="FileNotFoundException">    Thrown when the requested file is not present. </exception>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   The contents of the resource file or empty if the resource file is empty or not found. </returns>
+    public string ReadResourceFile( string resourceFileName );
+
+    /// <summary>   Gets resource file reader. </summary>
+    /// <remarks>   2025-04-15. <para>
+    /// Returning the stream reader allows rewinding. </para> </remarks>
+    /// <exception cref="FileNotFoundException">    Thrown when the requested file is not present. </exception>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <returns>   The resource file reader. </returns>
+    public StreamReader GetResourceFileReader( string resourceFileName );
+
+    #endregion
+}